Added an if statment in OutgoingMenu.java
This ifstatement is meant to check whether the person has been logged in yet, and if they have, alerts them of inputting a number not in the range of the menu.

Only problem with this method is the ability to fluctuate between options. I need to set up a method to connect the backend and send out menu sizes and place that into a data field in Client.java